requirements.txt
#backend
fastapi[standard]
python-dotenv
google-genai

#database
duckdb
pandas

#frontend
streamlit
requests